start new thread, sort the objects based on key, count the number of calls to fnMut closure
