+++ closures/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod shirtStock;
+pub mod closure_ops;
